PdlClient brukes i PdlHealthIndicator.kt
