+++ src/main/kotlin/no/nav/familie/ef/søknad/health/PdlHealthIndicator.kt
@@ -1,10 +1,10 @@
 package no.nav.familie.ef.søknad.health
 
-import no.nav.familie.ef.søknad.integration.PdlStsClient
+import no.nav.familie.ef.søknad.integration.PdlClient
 import no.nav.familie.http.health.AbstractHealthIndicator
 import org.springframework.stereotype.Component
 
 @Component
-internal class PdlHealthIndicator(pdlStsClient: PdlStsClient)
-    : AbstractHealthIndicator(pdlStsClient, "pdl.personinfo")
+internal class PdlHealthIndicator(pdlClient: PdlClient)
+    : AbstractHealthIndicator(pdlClient, "pdl.personinfo")
