+++ src/main/kotlin/no/nav/familie/ef/søknad/integration/PdlClient.kt
@@ -3,19 +3,20 @@ package no.nav.familie.ef.søknad.integration
 import no.nav.familie.ef.søknad.config.PdlConfig
 import no.nav.familie.ef.søknad.exception.PdlRequestException
 import no.nav.familie.ef.søknad.integration.dto.pdl.*
-import no.nav.familie.http.client.AbstractRestClient
+import no.nav.familie.http.client.AbstractPingableRestClient
 import no.nav.familie.http.sts.StsRestClient
 import org.springframework.beans.factory.annotation.Qualifier
 import org.springframework.http.HttpHeaders
 import org.springframework.stereotype.Component
 import org.springframework.web.client.RestOperations
+import java.net.URI
 
 
 @Component
 class PdlClient(val pdlConfig: PdlConfig,
                 @Qualifier("restKlientMedApiKey") restOperations: RestOperations,
                 private val stsRestClient: StsRestClient)
-    : AbstractRestClient(restOperations, "pdl.personinfo") {
+    : AbstractPingableRestClient(restOperations, "pdl.personinfo") {
 
     fun hentSøker(personIdent: String): PdlSøker {
         val pdlPersonRequest = PdlPersonRequest(variables = PdlPersonRequestVariables(personIdent),
@@ -65,6 +66,13 @@ class PdlClient(val pdlConfig: PdlConfig,
             add("Nav-Consumer-Token", "Bearer ${stsRestClient.systemOIDCToken}")
             add("Tema", "ENF")
         }
+    }
+
+    override val pingUri: URI
+        get() = pdlConfig.pdlUri
+
+    override fun ping() {
+        operations.optionsForAllow(pingUri)
     }
 
 }
